Append "Request" to API function names
For better distinction between the API functions and the SDK functions, especially deleteInstallation.
