+++ packages/installations/src/api/create-installation-request.test.ts
@@ -18,9 +18,10 @@
 import { FirebaseError } from '@firebase/util';
 import { expect } from 'chai';
 import { SinonStub, stub } from 'sinon';
+import { CreateInstallationResponse } from '../interfaces/api-response';
 import { AppConfig } from '../interfaces/app-config';
 import {
-  RegisteredInstallationEntry,
+  InProgressInstallationEntry,
   RequestStatus
 } from '../interfaces/installation-entry';
 import { compareHeaders } from '../testing/compare-headers';
@@ -28,57 +29,98 @@ import { getFakeAppConfig } from '../testing/get-fake-app';
 import '../testing/setup';
 import {
   INSTALLATIONS_API_URL,
-  INTERNAL_AUTH_VERSION
+  INTERNAL_AUTH_VERSION,
+  PACKAGE_VERSION
 } from '../util/constants';
 import { ErrorResponse } from './common';
-import { deleteInstallation } from './delete-installation';
+import { createInstallationRequest } from './create-installation-request';
 
-const FID = 'foreclosure-of-a-dream';
+const FID = 'defenders-of-the-faith';
 
-describe('deleteInstallation', () => {
+describe('createInstallationRequest', () => {
   let appConfig: AppConfig;
   let fetchSpy: SinonStub<[RequestInfo, RequestInit?], Promise<Response>>;
-  let registeredInstallationEntry: RegisteredInstallationEntry;
+  let inProgressInstallationEntry: InProgressInstallationEntry;
+  let response: CreateInstallationResponse;
 
   beforeEach(() => {
     appConfig = getFakeAppConfig();
 
-    registeredInstallationEntry = {
+    inProgressInstallationEntry = {
       fid: FID,
-      registrationStatus: RequestStatus.COMPLETED,
-      refreshToken: 'refreshToken',
-      authToken: {
-        requestStatus: RequestStatus.NOT_STARTED
-      }
+      registrationStatus: RequestStatus.IN_PROGRESS,
+      registrationTime: Date.now()
     };
 
+    response = {
+      refreshToken: 'refreshToken',
+      authToken: {
+        token:
+          'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCeKKF2QT4fwpMeJf36POk6yJV_adQssw5c',
+        expiresIn: '604800s'
+      },
+      fid: FID
+    };
     fetchSpy = stub(self, 'fetch');
   });
 
   describe('successful request', () => {
     beforeEach(() => {
-      fetchSpy.resolves(new Response());
+      fetchSpy.resolves(new Response(JSON.stringify(response)));
     });
 
-    it('calls the deleteInstallation server API with correct parameters', async () => {
+    it('registers a pending InstallationEntry', async () => {
+      const registeredInstallationEntry = await createInstallationRequest(
+        appConfig,
+        inProgressInstallationEntry
+      );
+      expect(registeredInstallationEntry.registrationStatus).to.equal(
+        RequestStatus.COMPLETED
+      );
+    });
+
+    it('calls the createInstallation server API with correct parameters', async () => {
       const expectedHeaders = new Headers({
         'Content-Type': 'application/json',
         Accept: 'application/json',
-        Authorization: `${INTERNAL_AUTH_VERSION} refreshToken`,
         'x-goog-api-key': 'apiKey'
       });
-      const expectedRequest: RequestInit = {
-        method: 'DELETE',
-        headers: expectedHeaders
+      const expectedBody = {
+        fid: FID,
+        authVersion: INTERNAL_AUTH_VERSION,
+        appId: appConfig.appId,
+        sdkVersion: PACKAGE_VERSION
       };
-      const expectedEndpoint = `${INSTALLATIONS_API_URL}/projects/projectId/installations/${FID}`;
+      const expectedRequest: RequestInit = {
+        method: 'POST',
+        headers: expectedHeaders,
+        body: JSON.stringify(expectedBody)
+      };
+      const expectedEndpoint = `${INSTALLATIONS_API_URL}/projects/projectId/installations`;
 
-      await deleteInstallation(appConfig, registeredInstallationEntry);
-
+      await createInstallationRequest(appConfig, inProgressInstallationEntry);
       expect(fetchSpy).to.be.calledOnceWith(expectedEndpoint, expectedRequest);
       const actualHeaders = fetchSpy.lastCall.lastArg.headers;
       compareHeaders(expectedHeaders, actualHeaders);
     });
+  });
+
+  it('returns the FID from the request if the response does not contain one', async () => {
+    response = {
+      refreshToken: 'refreshToken',
+      authToken: {
+        token:
+          'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCeKKF2QT4fwpMeJf36POk6yJV_adQssw5c',
+        expiresIn: '604800s'
+      }
+    };
+    fetchSpy.resolves(new Response(JSON.stringify(response)));
+
+    const registeredInstallationEntry = await createInstallationRequest(
+      appConfig,
+      inProgressInstallationEntry
+    );
+    expect(registeredInstallationEntry.fid).to.equal(FID);
   });
 
   describe('failed request', () => {
@@ -96,7 +138,7 @@ describe('deleteInstallation', () => {
       );
 
       await expect(
-        deleteInstallation(appConfig, registeredInstallationEntry)
+        createInstallationRequest(appConfig, inProgressInstallationEntry)
       ).to.be.rejectedWith(FirebaseError);
     });
 
@@ -112,10 +154,11 @@ describe('deleteInstallation', () => {
       fetchSpy
         .onCall(0)
         .resolves(new Response(JSON.stringify(errorResponse), { status: 500 }));
-      fetchSpy.onCall(1).resolves(new Response());
+      fetchSpy.onCall(1).resolves(new Response(JSON.stringify(response)));
 
-      await expect(deleteInstallation(appConfig, registeredInstallationEntry))
-        .to.be.fulfilled;
+      await expect(
+        createInstallationRequest(appConfig, inProgressInstallationEntry)
+      ).to.be.fulfilled;
       expect(fetchSpy).to.be.calledTwice;
     });
   });
