+++ packages/installations/src/api/delete-installation-request.test.ts
@@ -18,10 +18,9 @@
 import { FirebaseError } from '@firebase/util';
 import { expect } from 'chai';
 import { SinonStub, stub } from 'sinon';
-import { CreateInstallationResponse } from '../interfaces/api-response';
 import { AppConfig } from '../interfaces/app-config';
 import {
-  InProgressInstallationEntry,
+  RegisteredInstallationEntry,
   RequestStatus
 } from '../interfaces/installation-entry';
 import { compareHeaders } from '../testing/compare-headers';
@@ -29,98 +28,57 @@ import { getFakeAppConfig } from '../testing/get-fake-app';
 import '../testing/setup';
 import {
   INSTALLATIONS_API_URL,
-  INTERNAL_AUTH_VERSION,
-  PACKAGE_VERSION
+  INTERNAL_AUTH_VERSION
 } from '../util/constants';
 import { ErrorResponse } from './common';
-import { createInstallation } from './create-installation';
+import { deleteInstallationRequest } from './delete-installation-request';
 
-const FID = 'defenders-of-the-faith';
+const FID = 'foreclosure-of-a-dream';
 
-describe('createInstallation', () => {
+describe('deleteInstallationRequest', () => {
   let appConfig: AppConfig;
   let fetchSpy: SinonStub<[RequestInfo, RequestInit?], Promise<Response>>;
-  let inProgressInstallationEntry: InProgressInstallationEntry;
-  let response: CreateInstallationResponse;
+  let registeredInstallationEntry: RegisteredInstallationEntry;
 
   beforeEach(() => {
     appConfig = getFakeAppConfig();
 
-    inProgressInstallationEntry = {
+    registeredInstallationEntry = {
       fid: FID,
-      registrationStatus: RequestStatus.IN_PROGRESS,
-      registrationTime: Date.now()
-    };
-
-    response = {
+      registrationStatus: RequestStatus.COMPLETED,
       refreshToken: 'refreshToken',
       authToken: {
-        token:
-          'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCeKKF2QT4fwpMeJf36POk6yJV_adQssw5c',
-        expiresIn: '604800s'
-      },
-      fid: FID
+        requestStatus: RequestStatus.NOT_STARTED
+      }
     };
+
     fetchSpy = stub(self, 'fetch');
   });
 
   describe('successful request', () => {
     beforeEach(() => {
-      fetchSpy.resolves(new Response(JSON.stringify(response)));
+      fetchSpy.resolves(new Response());
     });
 
-    it('registers a pending InstallationEntry', async () => {
-      const registeredInstallationEntry = await createInstallation(
-        appConfig,
-        inProgressInstallationEntry
-      );
-      expect(registeredInstallationEntry.registrationStatus).to.equal(
-        RequestStatus.COMPLETED
-      );
-    });
-
-    it('calls the createInstallation server API with correct parameters', async () => {
+    it('calls the deleteInstallation server API with correct parameters', async () => {
       const expectedHeaders = new Headers({
         'Content-Type': 'application/json',
         Accept: 'application/json',
+        Authorization: `${INTERNAL_AUTH_VERSION} refreshToken`,
         'x-goog-api-key': 'apiKey'
       });
-      const expectedBody = {
-        fid: FID,
-        authVersion: INTERNAL_AUTH_VERSION,
-        appId: appConfig.appId,
-        sdkVersion: PACKAGE_VERSION
-      };
       const expectedRequest: RequestInit = {
-        method: 'POST',
-        headers: expectedHeaders,
-        body: JSON.stringify(expectedBody)
+        method: 'DELETE',
+        headers: expectedHeaders
       };
-      const expectedEndpoint = `${INSTALLATIONS_API_URL}/projects/projectId/installations`;
+      const expectedEndpoint = `${INSTALLATIONS_API_URL}/projects/projectId/installations/${FID}`;
 
-      await createInstallation(appConfig, inProgressInstallationEntry);
+      await deleteInstallationRequest(appConfig, registeredInstallationEntry);
+
       expect(fetchSpy).to.be.calledOnceWith(expectedEndpoint, expectedRequest);
       const actualHeaders = fetchSpy.lastCall.lastArg.headers;
       compareHeaders(expectedHeaders, actualHeaders);
     });
-  });
-
-  it('returns the FID from the request if the response does not contain one', async () => {
-    response = {
-      refreshToken: 'refreshToken',
-      authToken: {
-        token:
-          'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCeKKF2QT4fwpMeJf36POk6yJV_adQssw5c',
-        expiresIn: '604800s'
-      }
-    };
-    fetchSpy.resolves(new Response(JSON.stringify(response)));
-
-    const registeredInstallationEntry = await createInstallation(
-      appConfig,
-      inProgressInstallationEntry
-    );
-    expect(registeredInstallationEntry.fid).to.equal(FID);
   });
 
   describe('failed request', () => {
@@ -138,7 +96,7 @@ describe('createInstallation', () => {
       );
 
       await expect(
-        createInstallation(appConfig, inProgressInstallationEntry)
+        deleteInstallationRequest(appConfig, registeredInstallationEntry)
       ).to.be.rejectedWith(FirebaseError);
     });
 
@@ -154,10 +112,11 @@ describe('createInstallation', () => {
       fetchSpy
         .onCall(0)
         .resolves(new Response(JSON.stringify(errorResponse), { status: 500 }));
-      fetchSpy.onCall(1).resolves(new Response(JSON.stringify(response)));
+      fetchSpy.onCall(1).resolves(new Response());
 
-      await expect(createInstallation(appConfig, inProgressInstallationEntry))
-        .to.be.fulfilled;
+      await expect(
+        deleteInstallationRequest(appConfig, registeredInstallationEntry)
+      ).to.be.fulfilled;
       expect(fetchSpy).to.be.calledTwice;
     });
   });
